+++ library/src/main/java/ch/temparus/android/dialog/holder/ListViewHolder.java
@@ -21,7 +21,6 @@ public class ListViewHolder implements HolderAdapter<BaseAdapter>, AdapterView.O
     private ListView mListView;
     private BaseAdapter mAdapter;
     private OnHolderListener mHolderListener;
-    private View.OnKeyListener mKeyListener;
     private Dialog.State mState;
     private long mLastStateChange;
     private int mBackgroundColorResource;
@@ -41,6 +40,9 @@ public class ListViewHolder implements HolderAdapter<BaseAdapter>, AdapterView.O
         if (view == null) {
             return;
         }
+        if (view.getParent() != null) {
+            ((ViewGroup) view.getParent()).removeView(view);
+        }
         mHeaderContainer.addView(view);
     }
 
@@ -48,6 +50,9 @@ public class ListViewHolder implements HolderAdapter<BaseAdapter>, AdapterView.O
     public void addFooter(View view) {
         if (view == null) {
             return;
+        }
+        if (view.getParent() != null) {
+            ((ViewGroup) view.getParent()).removeView(view);
         }
         mFooterContainer.addView(view);
     }
@@ -88,15 +93,6 @@ public class ListViewHolder implements HolderAdapter<BaseAdapter>, AdapterView.O
                 mIsInterceptTouchEventDisallowed = (top < 0 && mState == Dialog.State.EXPANDED);
             }
         });
-        mListView.setOnKeyListener(new View.OnKeyListener() {
-            @Override
-            public boolean onKey(View v, int keyCode, KeyEvent event) {
-                if (mKeyListener == null) {
-                    throw new NullPointerException("keyListener should not be null");
-                }
-                return mKeyListener.onKey(v, keyCode, event);
-            }
-        });
         mListView.scrollTo(0, 0);
         mHeaderContainer = (FrameLayout) view.findViewById(R.id.t_dialog__header_container);
         mHeaderContainer.setBackgroundColor(backgroundColor);
@@ -108,11 +104,6 @@ public class ListViewHolder implements HolderAdapter<BaseAdapter>, AdapterView.O
     @Override
     public void setOnItemClickListener(OnHolderListener holderListener) {
         mHolderListener = holderListener;
-    }
-
-    @Override
-    public void setOnKeyListener(View.OnKeyListener keyListener) {
-        mKeyListener = keyListener;
     }
 
     @Override