+++ library/src/main/java/ch/temparus/android/dialog/holder/RecyclerViewHolder.java
@@ -27,7 +27,6 @@ public class RecyclerViewHolder implements HolderAdapter<BaseAdapter> {
     private RecyclerView.LayoutManager mLayoutManager;
     private BaseAdapter mAdapter;
     private OnHolderListener mHolderListener;
-    private View.OnKeyListener mKeyListener;
     private Dialog.State mState;
     private long mLastStateChange;
     private int mBackgroundColorResource;
@@ -49,6 +48,9 @@ public class RecyclerViewHolder implements HolderAdapter<BaseAdapter> {
         if (view == null) {
             return;
         }
+        if (view.getParent() != null) {
+            ((ViewGroup) view.getParent()).removeView(view);
+        }
         mHeaderContainer.addView(view);
     }
 
@@ -56,6 +58,9 @@ public class RecyclerViewHolder implements HolderAdapter<BaseAdapter> {
     public void addFooter(View view) {
         if (view == null) {
             return;
+        }
+        if (view.getParent() != null) {
+            ((ViewGroup) view.getParent()).removeView(view);
         }
         mFooterContainer.addView(view);
     }
@@ -151,15 +156,6 @@ public class RecyclerViewHolder implements HolderAdapter<BaseAdapter> {
                 mIsInterceptTouchEventDisallowed = (top < 0 && mState == Dialog.State.EXPANDED);
             }
         });
-        mRecyclerView.setOnKeyListener(new View.OnKeyListener() {
-            @Override
-            public boolean onKey(View v, int keyCode, KeyEvent event) {
-                if (mKeyListener == null) {
-                    throw new NullPointerException("keyListener should not be null");
-                }
-                return mKeyListener.onKey(v, keyCode, event);
-            }
-        });
         mHeaderContainer = (FrameLayout) view.findViewById(R.id.t_dialog__header_container);
         mHeaderContainer.setBackgroundColor(backgroundColor);
         mHeaderContainer.setOnTouchListener(new View.OnTouchListener() {
@@ -182,11 +178,6 @@ public class RecyclerViewHolder implements HolderAdapter<BaseAdapter> {
     @Override
     public void setOnItemClickListener(OnHolderListener holderListener) {
         mHolderListener = holderListener;
-    }
-
-    @Override
-    public void setOnKeyListener(View.OnKeyListener keyListener) {
-        mKeyListener = keyListener;
     }
 
     @Override