+++ library/src/main/java/ch/temparus/android/dialog/holder/ViewHolder.java
@@ -1,7 +1,6 @@
 package ch.temparus.android.dialog.holder;
 
 import android.support.annotation.ColorRes;
-import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -24,7 +23,6 @@ public class ViewHolder implements Holder {
 
     private ViewGroup mHeaderContainer;
     private ViewGroup mFooterContainer;
-    private View.OnKeyListener mKeyListener;
 
     private View mContentView;
     private int mViewResourceId = INVALID;
@@ -43,6 +41,9 @@ public class ViewHolder implements Holder {
         if (view == null) {
             return;
         }
+        if (view.getParent() != null) {
+            ((ViewGroup) view.getParent()).removeView(view);
+        }
         mHeaderContainer.addView(view);
     }
 
@@ -50,6 +51,9 @@ public class ViewHolder implements Holder {
     public void addFooter(View view) {
         if (view == null) {
             return;
+        }
+        if (view.getParent() != null) {
+            ((ViewGroup) view.getParent()).removeView(view);
         }
         mFooterContainer.addView(view);
     }
@@ -66,16 +70,6 @@ public class ViewHolder implements Holder {
         int backgroundColor = parent.getResources().getColor(mBackgroundColor);
         ViewGroup contentContainer = (ViewGroup) view.findViewById(R.id.view_container);
         contentContainer.setBackgroundColor(backgroundColor);
-        contentContainer.setOnKeyListener(new View.OnKeyListener() {
-            @Override
-            public boolean onKey(View v, int keyCode, KeyEvent
-                    event) {
-                if (mKeyListener == null) {
-                    throw new NullPointerException("OnKeyListener should not be null");
-                }
-                return mKeyListener.onKey(v, keyCode, event);
-            }
-        });
         addContent(inflater, parent, contentContainer);
         mHeaderContainer = (ViewGroup) view.findViewById(R.id.t_dialog__header_container);
         mHeaderContainer.setBackgroundColor(backgroundColor);
@@ -94,11 +88,6 @@ public class ViewHolder implements Holder {
             }
         }
         container.addView(mContentView);
-    }
-
-    @Override
-    public void setOnKeyListener(View.OnKeyListener keyListener) {
-        mKeyListener = keyListener;
     }
 
     @Override