+++ library/src/main/java/ch/temparus/android/dialog/listeners/OnCancelListener.java
@@ -2,16 +2,19 @@ package ch.temparus.android.dialog.listeners;
 
 import ch.temparus.android.dialog.Dialog;
 
+import java.io.Serializable;
+
 /**
  * Interface for listening to the cancel event of {@link Dialog}.
  *
  * @author Sandro Lutz
  */
-public interface OnCancelListener {
+public interface OnCancelListener extends Serializable {
 
     /**
      * Called when the dialog has been cancelled.
      * @param dialog dialog instance
+     * @return true - if dialog can be dismissed; false - otherwise
      */
-    void onCancel(Dialog dialog);
+    boolean onCancel(Dialog dialog);
 }