+++ library/src/main/java/ch/temparus/android/dialog/listeners/OnClickListener.java
@@ -3,12 +3,14 @@ package ch.temparus.android.dialog.listeners;
 import android.view.View;
 import ch.temparus.android.dialog.Dialog;
 
+import java.io.Serializable;
+
 /**
  * Interface for listening to all click events of the {@link Dialog} (including header and footer!).
  *
  * @author Sandro Lutz
  */
-public interface OnClickListener {
+public interface OnClickListener extends Serializable {
 
     /**
      * Called when a click event occurred on the {@link Dialog}.