+++ library/src/main/java/ch/temparus/android/dialog/listeners/OnDismissListener.java
@@ -2,12 +2,14 @@ package ch.temparus.android.dialog.listeners;
 
 import ch.temparus.android.dialog.Dialog;
 
+import java.io.Serializable;
+
 /**
  * Interface for listening to the dismiss event of {@link Dialog}.
  *
  * @author Sandro Lutz
  */
-public interface OnDismissListener {
+public interface OnDismissListener extends Serializable {
 
     /**
      * Called when the dialog has been dismissed.