+++ library/src/main/java/ch/temparus/android/dialog/listeners/OnStateChangeListener.java
@@ -2,12 +2,14 @@ package ch.temparus.android.dialog.listeners;
 
 import ch.temparus.android.dialog.Dialog;
 
+import java.io.Serializable;
+
 /**
  * Interface for listening to state changes in implementations of {@link ch.temparus.android.dialog.holder.Holder}.
  *
  * @author Sandro Lutz
  */
-public interface OnStateChangeListener {
+public interface OnStateChangeListener extends Serializable {
 
     /**
      * Called when dialog's state has changed.